Add sealed hierarchy analysis.
In general, members can be overridden outside the library. When
that is the case, they must preserve an *equivalent* result type.

Members that cannot be overridden outside the library can evolve
with a *subtype* in future versions. We can decide that property
with an analysis of final members and sealed class hierarchies.