Only track translated Binders in TypeTranslator.
`Type`s are not implicitly recursive. When they are recursive, they
are made explicitly so with `RecType`s. So we do not actually need
to track `Type`s in general for recursion.

What we really need is to track `Binders` so that we can translated
references to their parts (`ParamRef`s and `RecThis`es).